Corrige les warnings int/unsigned (-Wsign-conversion) dans tout le code BioSpring
27 emplacements corrigés dans le code propre au projet (le tiers vendorisé
src/IO/xdrfile/*.c, bibliothèque GROMACS, n'est pas touché). Pour chaque
cas, décision individuelle plutôt que cast systématique :

- Compteurs/index qui sont réellement toujours >= 0 sans dépendance à un
  sentinel négatif : type changé en unsigned/size_t (Spring::_id/getId/
  setId, désormais unsigned car _springs.size()-1 est toujours >= 0 et
  aucun code ne s'appuie sur une valeur négative, contrairement à
  Particle::_id).

- Valeurs qui DOIVENT rester signées, cast explicite au point de
  consommation unsigned :
  - Particle::_id/getId() : sert de sentinel "non assigné" dans
    SpringNetwork::isProbeParticle (getId() >= 0).
  - Reduce::getReduceRuleIdFromName() : -1 = "not found".
  - matrix.cpp rows_/cols_ : boucles décroissantes type
    `for (int i = rows_ - 2; i >= 0; --i)` qui débilement en cas
    d'unsigned (sous-flux catastrophique).
  - Boucles OpenMP `int i` : MSVC ne supporte que OpenMP 2.0, qui exige
    un compteur signé pour `#pragma omp parallel for`.
  - Interactor::_nbpositions : adresse passée directement à
    IIMD_send_coords(const int*, ...), ABI C externe fixe (MDDriver).
  - Champs miroir du schéma NetCDF sur disque (SpnBuffer::springs,
    nbofspringsperparticle, resids : tous nc_INT côté fichier).
  - Topology.hpp residue_id()/atom_id() : certains fichiers PDB utilisent
    une numérotation de résidu négative.

- Arithmétique d'itérateur (vector::begin() + index unsigned nécessite un
  difference_type signé) : cast ponctuel vers ptrdiff_t/size_t
  (SpringNetwork.h::removeInteractor, ParticleCollection.hpp::remove_particle).

Vérifié avec un build scratch clang -Wsign-conversion -Wsign-compare :
408 warnings avant, 0 après (hors xdrfile). Suite de tests complète
(251/251) verte après recompilation normale.

+++ src/IO/PDBReader.cpp
@@ -65,12 +65,12 @@ std::vector<std::pair<size_t, size_t>> PDBReader::parseConectLine(const std::str
 {
     std::vector<std::pair<size_t, size_t>> pairs_indexes;
 
-    size_t serial = std::stoi(line.substr(6, 5));
+    size_t serial = static_cast<size_t>(std::stoi(line.substr(6, 5)));
 
     // Helper function to try parsing and add to vector if successful
-    auto try_add_pair = [&](int start_index) {
+    auto try_add_pair = [&](size_t start_index) {
         try {
-            size_t connected_serial = std::stoi(line.substr(start_index, 5));
+            size_t connected_serial = static_cast<size_t>(std::stoi(line.substr(start_index, 5)));
             pairs_indexes.emplace_back(serial, connected_serial);
         } catch (const std::exception&) {
             // Catches std::invalid_argument and std::out_of_range
